# Faster AVX2 matrix multiplications for lgacy quants