--- a/llamafile/tinyblas_cpu_sgemm.inc
+++ b/llamafile/tinyblas_cpu_sgemm.inc
@@ -320,12 +320,11 @@ bool llamafile_sgemm(long m, long n, long k, const void *A, long lda, const void
                 return true;
             }
         }
-    }
-#endif
-    if ((Btype == GGML_TYPE_Q8_0 || Btype == GGML_TYPE_Q8_1) && Ctype == GGML_TYPE_F32) {
-        assert (QK8_0 == QK8_1 == QK4_0 == QK4_1 == QK5_0 == QK5_1 == 32);
-        if (iqk_mul_mat(m, n, k*QK8_0, Atype, A, B, (float *)C, ldc, ith, nth)) {
-            return true;
+        if ((Btype == GGML_TYPE_Q8_0 || Btype == GGML_TYPE_Q8_1) && Ctype == GGML_TYPE_F32) {
+            assert (QK8_0 == QK8_1 == QK4_0 == QK4_1 == QK5_0 == QK5_1 == 32);
+            if (iqk_mul_mat(m, n, k * QK8_0, Atype, A, B, (float *)C, ldc, ith, nth)) {
+                return true;
+            }
         }
     }
 #endif