@ikawrakow 2024-05-10
@jart I adapted to head. But to get back the Ryzen-7950X performance I had to make two separate `iqk_mul_map` versions (one for generic `AVX2` and one with `AVX512F+AVX512VNNI+AVX512VL` enabled).